lahaina_*QGKI: disable QTI_RPM_STATS_LOG
QTI_RPM_STATS_LOG adds support for logging various statistics related
to the QPM subsystem, such as the number of times each RPM command
was sent, the average and maximum response times of RPM commands, and
the number of times RPM requests were retried.

Keeping this enabled may add some overhead to the system, as it
requires additional kernel code to be executed to collect and process
the RPM statistics.

Change-Id: I844c89b4e6f2ade1fc0acd2fa395608788b68868
Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>